+++ packages/sdk/src/sync/subgraph/index.ts
@@ -1,7 +1,2 @@
 export { SubgraphSDKSyncAdapter } from "./adapter";
-export {
-  makeSubgraphQuery,
-  fetchLatestIndexedBlock,
-  entityIdFromBlockNumber,
-  entityIdWithEntityIndexFromBlockNumber,
-} from "./fetch";
+export * as SubgraphUtils from "./fetch";